(refactor): Changing structure, Dice is a Value Object
