+++ src/Domain/ValueObjects/Utils/Result.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace RPGPlayground\Domain\ValueObjects\App;
+namespace RPGPlayground\Domain\ValueObjects\Utils;
 
 /**
  * @template T
